% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{assets_download}
\alias{assets_download}
\title{assets download}
\usage{
assets_download(
  items,
  assets_name,
  output_dir = ".",
  overwrite = FALSE,
  items_max = Inf,
  progress = TRUE,
  ...,
  headers = character()
)
}
\arguments{
\item{items}{a \code{stac_item_collection} or \code{stac_item} object
representing the result of \code{/stac/search},
\code{/collections/{collectionId}/items} or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}

\item{assets_name}{a \code{character} with the assets names to be filtered.}

\item{output_dir}{a \code{character} directory in which the assets will be
saved.}

\item{overwrite}{a \code{logical} if TRUE will replaced the existing file,
if FALSE a warning message is shown.}

\item{progress}{a \code{logical} indicating if a progress bar must be
shown or not. Defaults to \code{TRUE}.}

\item{...}{other params to be passed to \link[httr]{GET} method.}

\item{headers}{a \code{character} of named arguments to be passed as
HTTP request headers.}

\item{items_size}{a \code{numeric} corresponding how many items will be
downloaded.}
}
\value{
The same \code{stac_item_collection} or \code{stac_item} object, with the
link of the item pointing to the directory where the assets were saved.
}
\description{
The \code{assets_download} function downloads the assets
provided by the STAC API.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/stac") \%>\%
  stac_search(collections = "CB4_64_16D_STK-1") \%>\%
  stac_search(limit = 2) \%>\%
  get_request() \%>\%
  assets_download(assets_name = "thumbnail", output_dir = "/home/felipe/",
  overwrite = FALSE)
}

}
\seealso{
\code{\link{stac_search}}, \code{\link{items}}, \code{\link{get_request}}
}
