% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-funs.R
\name{assets_filter}
\alias{assets_filter}
\alias{assets_filter.STACItemCollection}
\alias{assets_filter.STACItem}
\title{Assets filter}
\usage{
assets_filter(items, ..., filter_fn = NULL)

\method{assets_filter}{STACItemCollection}(items, ..., filter_fn = NULL)

\method{assets_filter}{STACItem}(items, ..., filter_fn = NULL)
}
\arguments{
\item{items}{a \code{STACItemCollection} object representing
the result of \verb{/stac/search}, \code{/collections/{collectionId}/items}.}

\item{...}{additional arguments. See details.}

\item{filter_fn}{a \code{function} that will be used to filter the
attributes listed in the properties.}
}
\value{
a \code{list} with the attributes of date, bands and paths.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
