% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-funs.R
\name{assets_function}
\alias{assets_function}
\alias{assets_download}
\alias{assets_download.STACItem}
\alias{assets_download.STACItemCollection}
\alias{assets_url}
\alias{assets_url.STACItem}
\alias{assets_url.STACItemCollection}
\alias{assets_url.default}
\alias{assets_select}
\alias{assets_select.STACItem}
\alias{assets_select.STACItemCollection}
\alias{assets_select.default}
\alias{assets_rename}
\alias{assets_rename.STACItem}
\alias{assets_rename.STACItemCollection}
\title{Assets functions}
\usage{
assets_download(
  items,
  asset_names = NULL,
  output_dir = getwd(),
  overwrite = FALSE,
  items_max = Inf,
  progress = TRUE,
  download_fn = NULL,
  ...,
  fn = deprecated()
)

\method{assets_download}{STACItem}(
  items,
  asset_names = NULL,
  output_dir = ".",
  overwrite = FALSE,
  items_max = Inf,
  progress = TRUE,
  download_fn = NULL,
  ...,
  fn = deprecated()
)

\method{assets_download}{STACItemCollection}(
  items,
  asset_names = NULL,
  output_dir = ".",
  overwrite = FALSE,
  items_max = Inf,
  progress = TRUE,
  download_fn = NULL,
  ...,
  fn = deprecated()
)

assets_url(items, asset_names = NULL, append_gdalvsi = TRUE)

\method{assets_url}{STACItem}(items, asset_names = NULL, append_gdalvsi = TRUE)

\method{assets_url}{STACItemCollection}(items, asset_names = NULL, append_gdalvsi = TRUE)

\method{assets_url}{default}(items, asset_names = NULL, append_gdalvsi = TRUE)

assets_select(items, asset_names = NULL, ..., select_fn = NULL)

\method{assets_select}{STACItem}(items, asset_names = NULL, ..., select_fn = NULL)

\method{assets_select}{STACItemCollection}(items, asset_names = NULL, ..., select_fn = NULL)

\method{assets_select}{default}(items, asset_names = NULL, ..., select_fn = NULL)

assets_rename(items, rename = NULL, names_fn = NULL, ...)

\method{assets_rename}{STACItem}(items, mapper = NULL, ...)

\method{assets_rename}{STACItemCollection}(items, mapper = NULL, ...)
}
\arguments{
\item{items}{a \code{STACItem} or \code{STACItemCollection} object
representing the result of \verb{/stac/search},
\code{/collections/{collectionId}/items} or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}

\item{asset_names}{a \code{character} vector with the assets names to be selected.}

\item{output_dir}{a \code{character} directory in which the assets will be
saved. Default is the working directory (\code{getwd()})}

\item{overwrite}{a \code{logical} if TRUE will replaced the existing file,
if FALSE a warning message is shown.}

\item{items_max}{a \code{numeric} corresponding how many items will be
downloaded.}

\item{progress}{a \code{logical} indicating if a progress bar must be
shown or not. Defaults to \code{TRUE}.}

\item{download_fn}{a \code{function} to handle the list of assets for each item.
Using this function you can change the hrefs for each asset, as well as use
another request verb, such as POST.}

\item{...}{config parameters to be passed to \link[httr:GET]{GET},
such as \link[httr:add_headers]{add_headers} or
\link[httr:set_cookies]{set_cookies}. Used in \code{assets_download} function.}

\item{fn}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
use \code{download_fn} parameter instead.}

\item{append_gdalvsi}{a \code{logical}  if true, gdal drivers are
included in the URL of each asset. The following schemes are supported:
HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).}

\item{names_fn}{a \code{function} used to produce assets names from an
asset object.}

\item{filter_fn}{a \code{function} that will be used to filter the
attributes listed in the properties.}

\item{keep_empty_items}{a \code{logical} indicating if empty items should be
kept.}
}
\value{
\itemize{
\item \code{assets_download()}: returns the same object as the provided item
(\code{STACItemCollection} or \code{STACItem}), however the \code{href} property points to
the location where the asset was downloaded

\item \code{assets_url()}: returns a character vector with \code{href} of each feature.

\item \code{assets_select()}: returns the same object as the provided item
(\code{STACItemCollection} or \code{STACItem}), however with the selected assets.

\item \code{assets_rename()}: returns the same object as the provided item
(\code{STACItemCollection} or \code{STACItem}), however with the assets renamed.
}
}
\description{
These functions provide support to work with \code{STACItemCollection} and
\code{STACItem} objects.

\itemize{
\item \code{assets_download()}: Downloads the assets provided by the STAC API.

\item \code{assets_url()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Returns a list
with href of each feature.
For the URL you can add the GDAL library drivers for the following schemes:
HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).

\item \code{assets_select()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Selects the
assets of each feature by its name.

\item \code{assets_rename()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Rename each
asset feature name by an expression or a function.
}
}
\examples{
\dontrun{
 # assets_download function
 stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
   stac_search(collections = "CB4_64_16D_STK-1",
               datetime = "2019-06-01/2019-08-01") \%>\%
   stac_search() \%>\%
   get_request() \%>\%
   assets_download(asset_names = "thumbnail", output_dir = tempdir())
}

\dontrun{
 # assets_url function
 stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
  stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
         datetime = "2017-08-01/2018-03-01",
         bbox = c(-48.206,-14.195,-45.067,-12.272)) \%>\%
  get_request() \%>\% items_fetch(progress = FALSE)

 stac_item \%>\% assets_url()
}

\dontrun{
 # assets_select function
 stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
  stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
         datetime = "2017-08-01/2018-03-01",
         bbox = c(-48.206,-14.195,-45.067,-12.272)) \%>\%
  get_request() \%>\% items_fetch(progress = FALSE)

 stac_item \%>\% assets_select(asset_names = "NDVI")
}

\dontrun{
items <- stac("https://planetarycomputer.microsoft.com/api/stac/v1") \%>\%
  stac_search(collections = c("landsat-8-c2-l2", "sentinel-2-l2a"),
              bbox = c(xmin = -64.85976089, ymin = -10.49199395,
                       xmax = -64.79272527, ymax =-10.44736091),
              datetime = "2019-01-01/2019-06-28",
              limit = 50) \%>\%
  post_request()

# Selects assets by name
items <- assets_select(items, c("B02", "B03", "SR_B1", "SR_B2"))
# Renames the landsat assets
items <- assets_rename(items,
                       SR_B1 = "blue",
                       SR_B2 = "green",
                       B02   = "blue",
                       B03   = "green")
# Get the assets url's
assets_url(items)
}

}
\seealso{
\code{\link[=stac_search]{stac_search()}}, \code{\link[=items]{items()}}, \code{\link[=get_request]{get_request()}}
}
