% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2-utils.R
\name{cql2_bbox_as_geojson}
\alias{cql2_bbox_as_geojson}
\title{CQL2 helper function}
\usage{
cql2_bbox_as_geojson(bbox)
}
\arguments{
\item{bbox}{A \code{numeric} containing a bbox with c(xmin, ymin, xmax, ymax).}
}
\value{
GeoJSON object
}
\description{
Function to convert bounding box (bbox) to a GeoJSON object
to be used as argument of CQL2 spatial operators.
}
\examples{
\dontrun{


}

}
