% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2.R
\name{cql2_functions}
\alias{cql2_functions}
\alias{cql2_json}
\alias{cql2_text}
\title{Convert R expressions to CQL2}
\usage{
cql2_json(expr)

cql2_text(expr)
}
\arguments{
\item{expr}{An R expression to be represented in CQL2}
}
\description{
These functions convert R expressions to CQL2 standard (\code{TEXT} or \code{JSON}).
}
