% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2-utils.R
\name{cql2_temporal}
\alias{cql2_temporal}
\alias{cql2_date}
\alias{cql2_timestamp}
\alias{cql2_interval}
\title{CQL2 temporal types}
\usage{
cql2_date(x)

cql2_timestamp(x)

cql2_interval(start = "..", end = "..")
}
\arguments{
\item{x, start, end}{\code{character} string containing valid \code{date} or \code{timestamp}.}
}
\value{
CQL2 temporal value
}
\description{
These are helper functions to create temporal literal values to be
passed into CQL2 expressions.
}
\examples{
\dontrun{


}

}
