% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-funs.R
\name{doc_query}
\alias{doc_query}
\title{Document utils functions}
\usage{
doc_query(d)
}
\arguments{
\item{d}{an \code{RSTACDocument} object}
}
\value{
a \code{RSTACQuery} object with the predecessor subclass with the
fields used in the request.
}
\description{
Document utils functions
}
\keyword{internal}
