% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{.build_stac}
\alias{.build_stac}
\title{stac object builder function}
\usage{
.build_stac(url, endpoint, params, subclass, base_stac)
}
\arguments{
\item{url}{a \code{character} informing the base url of a
STAC web service.}

\item{endpoint}{a \code{character} a path to be appended in the final
url.}

\item{params}{a named \code{list} with all url query parameters to be
appended in the url.}

\item{subclass}{a \code{character} corresponding to the subclass of the
object to be created.}

\item{base_stac}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{collections},
or \code{items} functions.}
}
\value{
A \code{stac} object with subclass \code{subclass} containing all request
 parameters to be provided to API service.
}
\description{
The \code{.build_stac} function builds a stac object based on a
 given \code{stac} object and others parameters.
}
\author{
Rolf Simoes
}
