% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{.check_response}
\alias{.check_response}
\title{STAC utils}
\usage{
.check_response(res, expected)
}
\arguments{
\item{res}{...}

\item{expected}{...}
}
\value{
...
}
\description{
The \code{stac_request} ...
}
\author{
Rolf Simoes
}
