% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac_utils.R
\name{.check_rfc}
\alias{.check_rfc}
\title{check date time as RFC patterns}
\usage{
.check_rfc(datetime)
}
\arguments{
\item{datetime}{Either a date-time or an interval, open or closed.
Date and time expressions adhere to RFC 3339. Open intervals are
expressed using double-dots.
Examples:
\itemize{
  \item A date-time: \code{"2018-02-12T23:20:50Z"}
  \item A closed interval: \code{"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"}
  \item Open intervals: \code{"2018-02-12T00:00:00Z/.."} or
    \code{"../2018-03-18T12:31:12Z"}
}}
}
\value{
A \code{logical} if TRUE the date time provided is correct,
otherwise not.
}
\description{
Auxiliary function to check that the provided date time follows
the standards of RFC 3339
}
\author{
Felipe Carvalho
}
