% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{.detect_version}
\alias{.detect_version}
\title{verify stac object}
\usage{
.detect_version(url, ...)
}
\arguments{
\item{url}{a \code{character} informing the base url of a
STAC web service.}

\item{...}{other params to be passed to \link[httr]{GET} or
\link[httr]{POST} methods.}
}
\value{
a \code{character}  with the STAC version found or an error if it is
 not possible to obtain it.
}
\description{
verify stac object
}
