% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{.file_ext}
\alias{.file_ext}
\title{Helper function of \code{assets_download}}
\usage{
.file_ext(asset_url)
}
\arguments{
\item{asset_url}{A \code{character} URL provided from a \code{stac_search}.}
}
\value{
A \code{character} of the extracted file extension.
}
\description{
The \code{.file_ext} is function to extract the extension
from a file
}
\author{
Implemented by \code{tools package}
}
