% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{.item_download}
\alias{.item_download}
\title{Helper function of \code{assets_download}}
\usage{
.item_download(stac_item, assets_name = c(), output_dir = "./")
}
\arguments{
\item{stac_item}{A  \code{stac_item} object expressing a STAC
search criteria provided by \code{stac_item} function.}

\item{assets_name}{A \code{character} with the assets names to be filtered.}

\item{output_dir}{A \code{character} directory in which the images will be
saved.}
}
\value{
The same \code{stac_item} object, but with the link of the item
 pointing to the directory where the assets were saved.
}
\description{
the \code{.item_download} function downloads the assets of a
 stac_item
}
