% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_url.R
\name{.make_url}
\alias{.make_url}
\title{make url}
\usage{
.make_url(url, endpoint = "/stac", params = list())
}
\arguments{
\item{url}{A \code{character} informing the base url of a
STAC web service.}

\item{endpoint}{A \code{character} ..}

\item{params}{A \code{list} ...}
}
\value{
A url ...
}
\description{
This function
}
\author{
Rolf Simoes
}
