% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac_utils.R
\name{.query_encode}
\alias{.query_encode}
\title{STAC utils}
\usage{
.query_encode(params)
}
\arguments{
\item{params}{...}
}
\value{
...
}
\description{
The \code{.query_encode} ...
}
\author{
Rolf Simoes
}
