% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{.select_assets}
\alias{.select_assets}
\title{Helper function of \code{assets_download} function}
\usage{
.select_assets(assets_list = list(), assets_names = c())
}
\arguments{
\item{assets_list}{A \code{list} with the information of each item provided
by API STAC}

\item{assets_names}{A \code{character} with the assets names to be filtered.}
}
\value{
A \code{list} in the same format as the list of assets, but with the
 selected assets names.
}
\description{
The helper function \code{.select_assets} selects the names of
each asset provided by users
}
\author{
Felipe Carvalho
}
