% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{.stac_method}
\alias{.stac_method}
\title{stac method}
\usage{
.stac_method(stac)
}
\arguments{
\item{stac}{A \code{stac} object expressing a STAC search criteria or
any \code{stac_*} object.}
}
\value{
An error if the given object does not belong to a \code{stac} class,
otherwise the http request method is added to the attributes.
}
\description{
checks if the \code{stac} object inherits the \code{stac} class
}
\author{
Rolf Simoes
}
