% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{.stac_request}
\alias{.stac_request}
\title{make url}
\usage{
.stac_request(
  url,
  endpoint = "/stac",
  params = list(),
  headers = list(),
  method = "get"
)
}
\arguments{
\item{url}{A \code{character} informing the base url of a
STAC web service.}

\item{endpoint}{A \code{character} ..}

\item{params}{A \code{list} ...}

\item{headers}{A \code{list} ...}

\item{method}{A \code{character} ...}
}
\value{
A res ...
}
\description{
This function
}
\author{
Rolf Simoes
}
