% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac_utils.R
\name{.verify_datetime}
\alias{.verify_datetime}
\title{verify date time}
\usage{
.verify_datetime(datetime)
}
\arguments{
\item{datetime}{Either a date-time or an interval, open or closed.
Date and time expressions adhere to RFC 3339. Open intervals are
expressed using double-dots.
Examples:
\itemize{
  \item A date-time: \code{"2018-02-12T23:20:50Z"}
  \item A closed interval: \code{"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"}
  \item Open intervals: \code{"2018-02-12T00:00:00Z/.."} or
    \code{"../2018-03-18T12:31:12Z"}
}}
}
\value{
An error if the date does not follow the specified standards or the
 date time provided as \code{character}.
}
\description{
Auxiliary function to check whether the date time follows
RFC 3339 standard.
}
\author{
Felipe Carvalho
}
