% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_filter.R
\name{ext_filter}
\alias{ext_filter}
\title{CQL2 filter extension}
\usage{
ext_filter(q, expr, lang = NULL, crs = NULL)
}
\description{
CQL2 is an OGC standard that enables complex filter expressions on OAFeat3
or STAC web services. CQL2 standard states that expressions can be
represented in JSON or TEXT formats.

TODO: add docs
}
