% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_query.R
\name{extension_query}
\alias{extension_query}
\title{Extension functions}
\usage{
extension_query(s, ...)
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{...}{entries with format \code{<field> <operator> <value>}.}
}
\value{
A \code{stac} object containing all request parameters to be
passed to \code{stac_search} function.
}
\description{
The \code{extension_query} function implements query API extension (v0.8.1)
that can be used in \code{/stac/search} endpoint using HTTP POST method
(\code{\link{post_request}}). It allows that additional fields and
operators can be passed to \code{\link{stac_search}} function (see examples).
Each filter entry is an expression formed by
\code{<field> <operator> <value>} terms, where \code{<field>} refers to
a valid items' property. Supported \code{<fields>} depends
on STAC API service implementation. The users must rely on service providers'
documentation to know which properties can be passed to do the search.

The \code{extension_query} function accepts the following \code{<operators>}
\itemize{
\item \code{==} corresponds to '\code{eq}'
\item \code{!=} corresponds to '\code{neq}'
\item \code{<} corresponds to '\code{lt}'
\item \code{<=} corresponds to '\code{lte}'
\item \code{>} corresponds to '\code{gt}'
\item \code{>=} corresponds to '\code{gte}'
\item \code{\%startsWith\%} corresponds to '\code{startsWith}' and implements
a string prefix search operator.
\item \code{\%endsWith\%} corresponds to '\code{endsWith}' and implements a
string suffix search operator.
\item \code{\%contains\%}: corresponds to '\code{contains}' and implements a
string infix search operator.
\item \code{\%in\%}: corresponds to '\code{in}' and implements a vector search
operator.
}
}
\examples{
\dontrun{

stac(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
  stac_search(collections = "CB4_64_16D_STK") \%>\%
  extension_query("bdc:tile" == "022024") \%>\%
  post_request()
}

}
\seealso{
\code{\link{stac_search}}, \code{\link{post_request}}
}
\author{
Rolf Simoes
}
