% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{get_doc_query}
\alias{get_doc_query}
\title{Document utils functions}
\usage{
get_doc_query(d)
}
\arguments{
\item{d}{\code{RSTACDocument} object}
}
\value{
a \code{RSTACQuery} object with the predecessor subclass with the
 fields used in the request.
}
\description{
Document utils functions
}
