% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subfields}
\alias{get_subfields}
\title{Utility functions}
\usage{
get_subfields(x, ..., fields = NULL, merge = TRUE)
}
\arguments{
\item{x}{a \code{list} that represents the subfields of a field of
the \code{RSTACDocument} object.}

\item{...}{a named way to provide fields names to get the subfields values
from the \code{RSTACDocument} objects.}

\item{fields}{a \code{character} with the names of the fields to get the
subfields values from the \code{RSTACDocument} objects.}

\item{merge}{a \code{logical} corresponds to the join of the subfields. If
true, the subfields are joined in a named list. If not, the subfields appear
in their respective index of the original list.}
}
\value{
a \code{list} with the values of the selected subfields.
}
\description{
These function retrieves the vvalues of subfields in \code{RSTACDocument}
 object.
}
