% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{http_request}
\alias{http_request}
\alias{.get_request}
\alias{.post_request}
\alias{.make_url}
\title{HTTP request}
\usage{
.get_request(stac, headers = list())

.post_request(stac, headers = list())

.make_url(url, endpoint = "", params = list())
}
\arguments{
\item{stac}{A \code{stac} object expressing a STAC search criteria.}

\item{headers}{A named \code{list} informing HTTP request headers to be
passed to \code{curl} package request functions.}

\item{url}{A \code{character} informing the base url of a
STAC web service.}

\item{endpoint}{A \code{character} a path to be appended in the final
url.}

\item{params}{A \code{list} with all url query parameters to be
appended in final url.}
}
\value{
A regular \code{list} object containing all HTTP parsed responses
returned by \code{curl} package.

\code{.make_url} returns an url to access STAC endpoints.
}
\description{
\code{.get_request} is a function that generates HTTP
GET requests. The object returned includes the following fields:
\itemize{
  \item \code{url}: \code{character} containing requested url
  \item \code{status_code}: \code{integer} indicating HTTP response status
  \item \code{type}: \code{character} informing response content type
  \item \code{headers}: \code{character} vector with all HTTP response
  headers
  \item \code{content}: \code{list} object representing JSON document
}
Calling rountines must parse this object to further return the result to
the user or raise error depending on the response content.

\code{.post_request} is a function that generates HTTP POST
requests. The object returned includes the same field of \code{.get_request}.

\code{.make_url} is a helper function to generate url. The returned
url is formed by appending \code{endpoint} at the end of base url
informed by \code{url} parameter. If \code{endpoint} has multiple elements
it will be collapsed using \code{'/'} character.

Note that \code{.make_url} function differs from standards of relative URI
path resolution (RFC 3986). Any existing path in base url
is maintained in the final url, and a simple string contatenation is made
whithout including any character separator. For this reason, this function
does not support the query and fragment URI components in the base url.
}
\author{
Rolf Simoes
}
