% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{items_assets}
\alias{items_assets}
\title{items function}
\usage{
items_assets(obj_stac)
}
\arguments{
\item{obj_stac}{A \code{stac} object expressing a STAC search criteria
provided by \code{stac_items} functions.}
}
\value{
...
}
\description{
baixar as imagens que vem do res e organizar em repo
}
\author{
Felipe Carvalho
}
