% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{items_assign_change}
\alias{items_assign_change}
\alias{items_sign}
\alias{items_sign.STACItemCollection}
\alias{items_sign.STACItem}
\title{Utility functions}
\usage{
items_sign(items, sign_fn = NULL, ...)

\method{items_sign}{STACItemCollection}(items, sign_fn = NULL)

\method{items_sign}{STACItem}(items, sign_fn = NULL)
}
\arguments{
\item{items}{a \code{STACItemCollection} object representing
the result of \verb{/stac/search}, \code{/collections/{collectionId}/items}.}

\item{sign_fn}{a \code{function} to assign each assets in STAC items.}
}
\value{
A \code{STACItemCollection} object with the signed assets according to the
supplied parameter function.
}
\description{
A utility function to create signatures in hrefs of each
asset in the STAC items. To create the signatures, the provided function
must return a function, being a factory function, so the manufactured
function will sign the hrefs returning one feature per interaction.
}
\examples{
\dontrun{

# Defining BDC token
Sys.setenv("BDC_ACCESS_KEY" = <your_bdc_access_key>)

# STACItem object
stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
 stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
        datetime = "2017-08-01/2018-03-01",
        bbox = c(-48.206,-14.195,-45.067,-12.272)) \%>\%
 get_request() \%>\% items_sign(sign_fn = sign_bdc)

}

}
