% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items-funs.R
\name{items_filter}
\alias{items_filter}
\title{Utility functions}
\usage{
items_filter(items, ..., fn = NULL)
}
\arguments{
\item{items}{a \code{STACItemCollection} object representing
the result of \verb{/stac/search}, \code{/collections/{collectionId}/items}.}

\item{...}{expressions used to filter items of a \code{STACItemCollection}
object.}

\item{fn}{a \code{function} that will be used to filter the attributes
listed in the properties.}
}
\value{
a \code{STACItemCollection} object with the filtered properties.
}
\description{
This function filters for the attributes contained in the STAC
properties.
}
\examples{
\donttest{
# STACItemCollection object
stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
 stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
        datetime = "2017-08-01/2018-03-01",
        bbox = c(-48.206,-14.195,-45.067,-12.272)) \%>\%
 get_request() \%>\% items_filter(`eo:cloud_cover` < 10)

# Example with AWS STAC
items <- stac("https://earth-search.aws.element84.com/v0") \%>\%
  stac_search(collections = "sentinel-s2-l2a-cogs",
              bbox = c(-48.206,-14.195,-45.067,-12.272),
              datetime = "2018-06-01/2018-06-30",
              limit = 500) \%>\%
  post_request()

  items \%>\%
   items_filter(fn = function(x) {x[["eo:cloud_cover"]] < 10})
}

}
