% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{items_functions}
\alias{items_functions}
\alias{items_length}
\alias{items_matched}
\alias{items_fetch}
\title{STACItemCollection functions}
\usage{
items_length(items)

items_matched(items)

items_fetch(items, ..., progress = TRUE)
}
\arguments{
\item{items}{a \code{STACItemCollection} object.}

\item{...}{config parameters to be passed to \link[httr]{GET} or
\link[httr]{POST} methods, such as \link[httr]{add_headers} or
\link[httr]{set_cookies}.}

\item{progress}{a \code{logical} indicating if a progress bar must be
shown or not. Defaults to \code{TRUE}.}
}
\value{
The \code{items_length()} returns an \code{integer} value.
The \code{items_matched()} returns an \code{integer} value.
If STAC web server does not support this extension, returns \code{NULL}.
The \code{items_fetch()} returns an \code{STACItemCollection} with all
matched items.
}
\description{
The \code{items_length()} function shows how many items there are in
the \code{STACItemCollection} object.
The \code{items_matched()} function shows how many items matched the
search criteria. It support \code{search:metadata} (v0.8.0) and
\code{context} (v0.9.0) STAC API extensions.
The \code{items_fetch()} function request all STAC Items through
pagination.
}
\examples{
\donttest{

x <-
  stac("http://brazildatacube.dpi.inpe.br/stac") \%>\%
  stac_search(collections = "MOD13Q1") \%>\%
  stac_search(bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) \%>\%
  stac_search(limit = 500) \%>\%
  get_request()

x \%>\% items_length()
x \%>\% items_matched()
x \%>\% items_fetch()

}

}
