% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_matched.R
\name{items_matched}
\alias{items_matched}
\title{Items functions}
\usage{
items_matched(items)
}
\arguments{
\item{items}{a \code{stac_items} object representing the request
results of \code{/stac/search}, \code{/collections/{collectionId}/items}, or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}
}
\value{
An \code{integer} value. If STAC web server does not support this
extension, returns \code{NULL}.
}
\description{
The \code{items_matched} returns how many items matched the
search criteria. It implements \code{search:metadata} STAC API
extension (v0.8.0).
}
\examples{
\dontrun{

stac_search("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    get_request() \%>\%
    items_matched()
}

}
\author{
Rolf Simoes
}
