% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{post_request}
\alias{post_request}
\title{STAC functions}
\usage{
post_request(s, encode = c("multipart", "form", "json"), headers = c())
}
\arguments{
\item{s}{A \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{encode}{A \code{character} informing the request body
Content-Type. Accepted types \code{'json'} \code{('application/json')},
\code{'form'} \code{('application/x-www-form-urlencoded')},
and \code{'multipart'} \code{('multipart/form-data')}.}

\item{headers}{A \code{character} of named arguments to be passed as
HTTP request headers.}
}
\value{
Either a \code{stac_collection} or a \code{stac_items} object
depending of the \code{s} parameter.
}
\description{
The \code{post_request} is function that makes HTTP
requests to STAC web services, retrieves, and parse the data.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0", limit = 100) \%>\%
    post_request()
}

}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{stac_collections}}
\code{\link{stac_items}}
}
\author{
Rolf Simoes
}
