% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-utils.R
\name{preview_plot}
\alias{preview_plot}
\title{Plot preview images}
\usage{
preview_plot(url)
}
\arguments{
\item{url}{image URL to be plotted.}
}
\value{
A rastergrob grob from package \code{grid}.
}
\description{
This is a helper function to plot preview assets
(e.g. quicklook, thumbnail, rendered_preview).
Currently, only png, jpeg and jpg formats are supported.
}
