% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-funs.R
\name{RSTACDocument}
\alias{RSTACDocument}
\title{Document development functions}
\usage{
RSTACDocument(content, q, subclass)
}
\arguments{
\item{content}{a \code{list} data structure representing the JSON file
received in HTTP response (see \code{\link[=content_response]{content_response()}} function)}

\item{q}{a \code{RSTACQuery} object expressing the STAC query used
to retrieve the document.}

\item{subclass}{a \code{character} corresponding to the subclass of the
document to be created.}
}
\value{
The \code{RSTACDocument()} function returns a \code{RSTACDocument} object
with subclass defined by \code{subclass} parameter.
}
\description{
Document development functions
}
\keyword{internal}
