% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_s}
\alias{search_s}
\title{Endpoint functions}
\usage{
search_s(s, collections, ids, bbox, datetime, intersects, limit, ...)
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{collections}{a \code{character} vector of collection IDs to include in
the search for items. Only items in one of the provided collections will be
searched.}

\item{ids}{a \code{character} vector with item IDs. All other filter
parameters that futher restrict the number of search results are ignored.}

\item{bbox}{only features that have a geometry that intersects the
bounding box are selected. The bounding box is provided as four or six
numbers, depending on whether the coordinate reference system includes a
vertical axis (elevation or depth): \itemize{ \item Lower left corner,
coordinate axis 1 \item Lower left corner, coordinate axis 2 \item Lower
left corner, coordinate axis 3 (optional) \item Upper right corner,
coordinate axis 1 \item Upper right corner, coordinate axis 2 \item Upper
 right corner, coordinate axis 3 (optional) }

The coordinate reference system of the values is WGS84 longitude/latitude
(\url{http://www.opengis.net/def/crs/OGC/1.3/CRS84}). The values are in
most cases the sequence of minimum longitude, minimum latitude, maximum
longitude and maximum latitude. However, in cases where the box spans the
antimeridian the first value (west-most box edge) is larger than the third
value (east-most box edge).}

\item{datetime}{either a date-time or an interval. Date and time strings
needs to conform RFC 3339. Intervals are expressed by separating two
date-time strings by \code{'/'} character. Open intervals are expressed by
using \code{'..'} in place of date-time.

Examples: \itemize{ \item A date-time: \code{"2018-02-12T23:20:50Z"} \item
A closed interval: \code{"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"} \item
Open intervals: \code{"2018-02-12T00:00:00Z/.."} or
\code{"../2018-03-18T12:31:12Z"} }

Only features that have a \code{datetime} property that intersects the
interval or date-time informed in \code{datetime} are selected.}

\item{intersects}{a \code{character} value expressing GeoJSON geometries
objects as specified in RFC 7946. Only returns items that intersect with
the provided polygon.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed it defaults to the service implementation.}

\item{...}{any additional non standard filter parameter.}

\item{query}{a \code{stac_query} object representing extra search
fields. This parameter can be obtained by \code{\link{ext_query}} function.}
}
\value{
A \code{stac} object containing all request parameters to be provided
to \code{stac_request}.
}
\description{
(This document is based on STAC specification documentation
\url{https://github.com/radiantearth/stac-spec/blob/v0.8.0/api-spec/STAC.yaml}
and reproduces some of its parts)

The \code{stac_search} function implements \code{/stac/search} API endpoint
(v0.8.0). It prepares query parameters used in search API request, a
\code{stac} object with all filter parameters to be provided to
\code{stac_request}. The GeoJSON content returned by the \code{stac_request}
is a \code{stac_items} object, a regular R \code{list} representing a STAC
ItemCollection.
}
\examples{
\dontrun{
# GET request
stac_search(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
            collections = "MOD13Q1",
            limit = 1) \%>\%
     get_request()

# POST request
stac_search(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
            collections = "MOD13Q1") \%>\%
     post_request()
}

}
\seealso{
\code{\link{stac}}, \code{\link{ext_query}},
\code{\link{get_request}}, \code{\link{post_request}}
}
\author{
Rolf Simoes
}
