% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{sign_bdc}
\alias{sign_bdc}
\title{Signature in hrefs provided by the STAC from the Brazil Data Cube
 project.}
\usage{
sign_bdc(url, env = NULL, ...)
}
\arguments{
\item{url}{a \code{character} with the reference href link provided by
objects \code{STACItem} or \code{STACItemCollection}.}

\item{env}{a \code{enviroment} object created to store local variables of the
package.}

\item{...}{additional parameters can be supplied to the \code{GET} function
of the \code{httr} package.}
}
\value{
a \code{character} with the href signed by the token.
}
\description{
To sign the hrefs with your token you need to store it in an
environment variable. Use the environment variable \code{BDC_ACCESS_KEY}.
}
