% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{sign_planetary_computer}
\alias{sign_planetary_computer}
\title{Signature in hrefs provided by the STAC from Microsoft's Planetary
Computer.}
\usage{
sign_planetary_computer(...)
}
\arguments{
\item{...}{additional parameters can be supplied to the \code{GET} function
of the \code{httr} package.}
}
\value{
a \code{function} that signs each item assets.
}
\description{
To perform the signing of the hrefs a request is sent to
Planetary Computer servers and the returned content corresponds to the
token that will be used in the href.
}
