% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_collections.R
\name{stac_collections}
\alias{stac_collections}
\title{Endpoint functions}
\usage{
stac_collections(url, collection_id)
}
\arguments{
\item{url}{a \code{character} informing the base url of a
STAC web service.}

\item{collection_id}{a \code{character} collection id to be retrieved.}
}
\value{
If no \code{collection_id} is informed, \code{stac_collections} returns a
list of STAC Collections. Otherwise, it will return a \code{stac_collection} object
representing a specific STAC Collection.
}
\description{
The \code{stac_collections} function implements the WFS3 \code{/collections}
and \code{/collections/\{collectionId\}} endpoints (v0.8.0).

Each endpoint retrieves specific STAC objects:
\itemize{
  \item \code{/collections}: Returns a list of STAC Collection published in
    the STAC service
  \item \code{/collections/\{collectionId\}}: Returns a single STAC
    Collection object
}
}
\examples{
\dontrun{

stac_collections("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    stac_request()

stac_collections("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
                 collection_id = "CB_64_16D_STK") \%>\%
    stac_request()
}

}
\seealso{
\code{\link{get_request}}, \code{\link{post_request}},
 \code{\link{stac_items}}
}
\author{
Rolf Simoes
}
