% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_download.R
\name{stac_download}
\alias{stac_download}
\title{stac download}
\usage{
stac_download(res, output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
baixar as imagens que vem do res e organizar em repo
}
\author{
...
}
