% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_dryrun.R
\name{stac_dryrun}
\alias{stac_dryrun}
\title{stac functions}
\usage{
stac_dryrun(expr)
}
\arguments{
\item{expr}{A stac command expression to be evaluated.}
}
\value{
A \code{character} string containing the http URL request of the
stac command. If no \code{expr} is informed, returns \code{FALSE}.
}
\description{
This function returns the URL requests that would be
performed by stac commands without actually running them.
}
\examples{
stac_dryrun(stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0"))
stac_dryrun({
  stac_search(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
              collections = "MOD13Q1",
              bbox = c(-55.16335, -4.26325, -49.31739, -1.18355))
            })

}
\author{
Rolf Simoes
}
