% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-funs.R
\name{stac_functions}
\alias{stac_functions}
\alias{stac_version}
\alias{stac_type}
\title{Utility functions}
\usage{
stac_version(x, ...)

stac_type(x)
}
\arguments{
\item{x}{either a \code{rstac_query} object expressing a STAC query
criteria or any \code{rstac_doc}.}

\item{...}{config parameters to be passed to \link[httr:GET]{GET}
method, such as \link[httr:add_headers]{add_headers} or \link[httr:set_cookies]{set_cookies}.}
}
\value{
The \code{stac_version()} function returns a \code{character} STAC API
version.
}
\description{
These function retrieves information about either \code{rstac} queries
(\code{rstac_query} objects) or \code{rstac} documents
(\code{rstac_doc} objects).
}
