% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_items.R
\name{stac_items}
\alias{stac_items}
\title{Endpoint functions}
\usage{
stac_items(url, collection_id, item_id, ..., .limit, .next)
}
\arguments{
\item{url}{A \code{character} informing the base url of a
STAC web service.}

\item{collection_id}{A \code{character} with a collection id to retrieve
collection details.}

\item{item_id}{A \code{character} with item id to be fetched.
Only works if the \code{collection_id} is informed. This is equivalent to
the endpoint \code{/collections/\{collectionId\}/items/\{itemId\}}.}

\item{...}{Filter parameters. Accept the same filter parameters
of \code{\link{stac_search}} function.}

\item{.limit}{An \code{integer} defining the maximum number of results
to return. Defaults to 10.}

\item{.next}{An \code{integer} informing which set of results
to return. Values less than 1 means all pages will be retrieved.}
}
\value{
A \code{stac} object containing all request parameters to be
provided to \code{stac_request}.
}
\description{
The \code{stac_items} function implements WFS3
\code{/collections/\{collectionId\}/items}, and
\code{/collections/\{collectionId\}/items/\{itemId\}}
endpoints (v0.8.0).

Each endpoint retrieves specific STAC objects:
\itemize{
  \item \code{/collections/\{collectionId\}/items}: Returns a STAC Items
    collection (GeoJSON)
  \item \code{/collections/\{collectionId\}/items/\{itemId\}}: Returns a
    STAC Item (GeoJSON Feature)
}

The endpoint \code{/collections/\{collectionId\}/items} accepts the same
filters parameters of \code{\link{stac_search}} function.
}
\examples{
\dontrun{

stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "CB_64_16D_STK",
           bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) \%>\%
    stac_request()
stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "MOD13Q1",
           item_id = "MOD13Q1.A2019241.h13v09.006.2019262164754") \%>\%
    stac_request()
}

}
\seealso{
\code{\link{stac_request}}, \code{\link{stac_collections}}
}
\author{
Rolf Simoes
}
