% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{params_get_request}
\alias{params_get_request}
\alias{params_post_request}
\alias{content_get_response}
\alias{content_post_response}
\title{Generic methods of the stac object}
\usage{
params_get_request(s)

params_post_request(s, enctype)

content_get_response(s, res)

content_post_response(s, res, enctype)
}
\description{
STAC objects implement four generic S3 methods, such methods are
\code{params} and \code{response}. The \code{params} method defines and
manages the parameters before HTTP requests and are implemented for each
object created in the package, its basis is \code{params_get_request} for
GET and \code{params_post_request} for POST. The methods \code{response}
manage the types of documents returned by the response, its base is
\code{content_get_response} for GET and \code{content_post_response} for
POST.
}
