% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subfields}
\alias{subfields}
\title{Utility functions}
\usage{
subfields(x)
}
\arguments{
\item{x}{a \code{list} that represents the subfields of a field of
the \code{RSTACDocument} object.}
}
\value{
a \code{character} with the names of the subfields.
}
\description{
These function retrieves the names of subfields in \code{RSTACDocument}
 object.
}
