% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{assets_download}
\alias{assets_download}
\title{stac download}
\usage{
assets_download(res, output_dir = "./", curl_header = list())
}
\arguments{
\item{res}{res parameter}

\item{output_dir}{output_dir parameter}

\item{curl_header}{curl_he}
}
\description{
baixar as imagens que vem do res e organizar em repo
}
\author{
...
}
