% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{assets_download}
\alias{assets_download}
\title{stac download}
\usage{
assets_download(
  res,
  output_dir = "./",
  curl_hearder = list(),
  assets_name = c()
)
}
\arguments{
\item{res}{...}

\item{output_dir}{...}

\item{assets_name}{...}

\item{curl_header}{...}
}
\value{
...
}
\description{
The \code{assets_download} function downloads the assets
provided by the STAC API
}
\author{
Felipe Carvalho and Rolf Simoes
}
