% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{assets_download}
\alias{assets_download}
\title{assets download}
\usage{
assets_download(
  items,
  assets_name,
  output_dir = ".",
  progress = TRUE,
  headers = c()
)
}
\arguments{
\item{items}{a \code{stac_items} or \code{stac_item} object
representing the result of \code{/stac/search},
\code{/collections/{collectionId}/items} or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}

\item{assets_name}{a \code{character} with the assets names to be filtered.}

\item{output_dir}{a \code{character} directory in which the images will be
saved.}

\item{progress}{a \code{logical} indicating if a progress bar must be
shown or not. Defaults to \code{TRUE}.}

\item{headers}{a \code{character} of named arguments to be passed as
HTTP request headers.}
}
\value{
The same \code{stac_items} or \code{stac_item} object, with the
link of the item pointing to the directory where the assets were saved.
}
\description{
The \code{assets_download} function downloads the assets
provided by the STAC API.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
 stac_search(collections = "MOD13Q1",
            bbox = c(-55.16335, -4.26325, -49.31739, -1.18355),
            limit = 2) \%>\%
 get_request() \%>\%
 assets_download(assets_name = c("thumbnail"), output_dir = "~")
}

}
\seealso{
\code{\link{stac_search}}, \code{\link{get_request}},
 \code{\link{post_request}}
}
\author{
Felipe Carvalho and Rolf Simoes
}
