% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-funs.R
\name{assets_function}
\alias{assets_function}
\alias{assets_append_gdalvsi}
\alias{assets_select}
\alias{assets_select.STACItemCollection}
\alias{assets_select.STACItem}
\title{Assets functions}
\usage{
assets_append_gdalvsi(
  items,
  asset_names = NULL,
  sort = deprecated(),
  gdal_vsi_resolution = deprecated()
)

assets_select(items, asset_names)

\method{assets_select}{STACItemCollection}(items, asset_names = NULL, filter_fn = NULL)

\method{assets_select}{STACItem}(items, asset_names = NULL, filter_fn = NULL)
}
\arguments{
\item{items}{a \code{STACItemCollection} object representing
the result of \verb{/stac/search}, \code{/collections/{collectionId}/items}.}

\item{asset_names}{a \code{character} with the assets names to be
filtered. If \code{NULL} (default) all assets will be returned..}

\item{sort}{a \code{logical} if true the dates will be sorted
in increasing order. By default, the dates are sorted.}

\item{gdal_vsi_resolution}{a \code{logical}  if true, gdal drivers are
included in the URL of each asset. The following schemes are supported:
HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).}

\item{filter_fn}{a \code{function} that will be used to filter the
attributes listed in the properties.}

\item{...}{additional arguments. See details.}

\item{fn}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
use \code{filter_fn} parameter instead.}
}
\value{
a \code{list} with the attributes of date, bands and paths.
}
\description{
This function returns the \code{date}, \code{band} and
\code{URL} fields for each assets of an \code{STACItemCollection} object.
For the URL you can add the GDAL library drivers for the following schemes:
HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).
}
\examples{
\donttest{
# STACItemCollection object
stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
 stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
        datetime = "2017-08-01/2018-03-01",
        bbox = c(-48.206,-14.195,-45.067,-12.272)) \%>\%
 get_request() \%>\% items_fetch(progress = FALSE)

stac_item \%>\% assets_append_gdalvsi(asset_names = c("EVI", "NDVI"))
}

}
