% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-funs.R
\name{assets_list}
\alias{assets_list}
\title{STACItemCollection function}
\usage{
assets_list(
  items,
  asset_names = NULL,
  sort = TRUE,
  gdal_vsi_resolution = TRUE,
  assets_names = deprecated()
)
}
\arguments{
\item{items}{a \code{STACItemCollection} or \code{STACItem} object.}

\item{asset_names}{a \code{character} with the assets names to be
filtered. If \code{NULL} (default) all assets will be returned..}

\item{sort}{a \code{logical} if true the dates will be sorted
in increasing order. By default, the dates are sorted.}

\item{gdal_vsi_resolution}{a \code{logical}  if true, gdal drivers are
included in the URL of each asset. The following schemes are supported:
HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).}

\item{assets_names}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
use \code{asset_names} parameter instead.}
}
\value{
a items \code{STACItemCollection} or \code{STACItem} object with gdal virtual
file system added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link{assets_gdalvfs}()} function
instead.
}
\examples{
\dontrun{

x <- stac("http://brazildatacube.dpi.inpe.br/stac") \%>\%
  stac_search(collections = "CB4_64_16D_STK-1") \%>\%
  stac_search() \%>\%
  get_request()

x \%>\% assets_append_gdalvfs()
}

}
