% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Endpoint functions}
\usage{
collections(s, collection_id)
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{collection_id}{a \code{character} collection id to be retrieved.}
}
\value{
If no \code{collection_id} is informed, \code{stac_collections} returns a
list of STAC Collections. Otherwise, it will return a \code{stac_collection} object
representing a specific STAC Collection.
}
\description{
The \code{stac_collections} function implements the WFS3 \code{/collections}
and \code{/collections/\{collectionId\}} endpoints (v0.8.0).

Each endpoint retrieves specific STAC objects:
\itemize{
  \item \code{/collections}: Returns a list of STAC Collection published in
    the STAC service
  \item \code{/collections/\{collectionId\}}: Returns a single STAC
    Collection object
}
}
\examples{
\dontrun{

 stac_collections("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    get_request()

 stac_collections("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
                 collection_id = "MOD13Q1") \%>\%
    get_request()
}

}
\seealso{
\code{\link{get_request}}, \code{\link{post_request}},
 \code{\link{stac_items}}
}
\author{
Rolf Simoes
}
