% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Endpoint functions}
\usage{
collections(s, collection_id)
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac} or \code{collections} functions.}

\item{collection_id}{a \code{character} collection id to be retrieved.}
}
\value{
A \code{stac} object containing all search field parameters to be provided
to STAC API web service.
}
\description{
The \code{collections} function implements the WFS3 \code{/collections}
 and \code{/collections/\{collectionId\}} endpoints
 (v0.8.0, v0.8.1 and v0.9.0).

Each endpoint retrieves specific STAC objects:
\itemize{
  \item \code{/collections}: Returns a list of STAC Collection published in
    the STAC service
  \item \code{/collections/\{collectionId\}}: Returns a single STAC
    Collection object
}
}
\examples{
\donttest{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.1",
    force_version = "0.8.1") \%>\%
 collections() \%>\%
 get_request()

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.1",
      force_version = "0.8.1") \%>\%
collections(collection_id = "MOD13Q1") \%>\%
get_request()
}
}
\seealso{
\code{\link{get_request}}, \code{\link{post_request}},
 \code{\link{items}}
}
