% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2.R
\name{cql2}
\alias{cql2}
\title{Convert R expressions to CQL2}
\usage{
cql2(expr, lang = NULL, crs = NULL, env_expr = environment())
}
\arguments{
\item{expr}{An R expression to be represented in CQL2}

\item{lang}{a character with the syntax used in the filter.
It can be used in text format \code{cql2-text} or in JSON format
\code{cql2-json}. By default, \code{cql2-text} is used in \code{GET}
requests and \code{cql2-json} in \code{POST} requests.}

\item{crs}{a character with coordinate reference systems.
By default WGS84 is used, this parameter will rarely be used.}

\item{env}{a environment to evaluate the expression}
}
\description{
These functions convert R expressions to CQL2 standard (text or JSON).
}
