% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{.check_response}
\alias{.check_response}
\title{STAC utils}
\usage{
.check_response(res, expected)
}
\arguments{
\item{res}{a \code{httr} \code{response} object.}

\item{expected}{a \code{list} containing the expected parameters values.}
}
\value{
a \code{character} with document class
}
\description{
The \code{.check_response} function that checks if the request's
response is in accordance with the \code{expected} parameters.
}
\author{
Rolf Simoes
}
