% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac_utils.R
\name{.make_url}
\alias{.make_url}
\title{STAC utils}
\usage{
.make_url(url, endpoint = "", params = list())
}
\arguments{
\item{url}{a \code{character} informing the base url of a
STAC web service.}

\item{endpoint}{a \code{character} a path to be appended in the final
url.}

\item{params}{a named \code{list} with all url query parameters to be
appended in the url.}
}
\value{
\code{.make_url} returns an url to access STAC endpoints.
}
\description{
\code{.make_url} is a helper function to generate url. The returned
url is formed by appending \code{endpoint} at the end of base url
informed by \code{url} parameter. If \code{endpoint} has multiple elements
it will be collapsed using \code{'/'} character.

Note that \code{.make_url} function differs from standards of relative URI
path resolution (RFC 3986). Any existing path in base url
is maintained in the final url, and a simple string contatenation is made
whithout including any character separator. For this reason, this function
does not support the query and fragment URI components in the base url.
}
