% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download.R
\name{items_assets}
\alias{items_assets}
\title{items function}
\usage{
items_assets(obj_stac)
}
\arguments{
\item{obj_stac}{A \code{stac} object expressing a STAC search criteria
provided by \code{stac_items} functions.}
}
\value{
A \code{list} with information of the assets of each item, where
each index represents one item.
}
\description{
The \code{items_assets} function lists the names of each assets
from each STAC item.
}
\examples{
\dontrun{

obj_stac <- stac_search(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
            collections = "MOD13Q1",
            bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) \%>\%
stac_request()

items_names <- items_assets(obj_stac)
}

}
\author{
Felipe Carvalho and Rolf Simoes
}
