% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_fetch.R
\name{items_fetch}
\alias{items_fetch}
\title{Items function}
\usage{
items_fetch(items, progress = TRUE, headers = list())
}
\arguments{
\item{items}{A \code{stac_items} object representing the result of
\code{/stac/search}, \code{/collections/{collectionId}/items}, or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}

\item{progress}{A \code{logical} indicating if a progress bar must be
shown or not. Defaults to \code{TRUE}.}

\item{headers}{A \code{list} of named arguments to be passed as
http request headers. This is used in \emph{addition} to eventual headers
defined in \code{stac} object parameter.}
}
\value{
A \code{stac_items} object.
}
\description{
The \code{items_fetch} function returns the pagination of all
 items of the stac object
}
\examples{
\dontrun{

stac_search(url = "http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
            collections = "MOD13Q1",
            bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) \%>\%
    stac_request() \%>\%
    items_fetch()
}

}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{stac_collections}}
\code{\link{stac_items}}
}
\author{
Rolf Simoes
}
