% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_length.R
\name{items_length}
\alias{items_length}
\title{Items functions}
\usage{
items_length(items)
}
\arguments{
\item{items}{a \code{stac_items} object representing the request
results of \code{/stac/search}, \code{/collections/{collectionId}/items}, or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}
}
\value{
An \code{integer} value.
}
\description{
The \code{items_length} function returns how many items
there are in the \code{items} object.
}
\examples{
\dontrun{

stac_search("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    stac_request() \%>\%
    items_length()
}

}
\author{
Rolf Simoes
}
