% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_matched.R
\name{items_matched}
\alias{items_matched}
\title{Items functions}
\usage{
items_matched(items)
}
\arguments{
\item{items}{a \code{stac_item_collection} object representing the
request results of \code{/stac/search},
\code{/collections/{collectionId}/items}, or
\code{/collections/{collectionId}/items/{itemId}} endpoints.}
}
\value{
An \code{integer} value. If STAC web server does not support this
extension, returns \code{NULL}.
}
\description{
The \code{items_matched} returns how many items matched the
search criteria. It implements \code{search:metadata} STAC API
extension (v0.8.0).
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.1",
     force_version = "0.8.1") \%>\%
 stac_search(collections = "MOD13Q1",
         bbox = c(-55.16335, -4.26325, -49.31739, -1.18355),
         limit = 500) \%>\%
 get_request() \%>\%
 items_matched()

}

}
