% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.RSTACQuery}
\alias{print.STACCatalog}
\alias{print.STACCollectionList}
\alias{print.STACCollection}
\alias{print.STACItemCollection}
\alias{print.STACItem}
\title{Printing functions}
\usage{
\method{print}{RSTACQuery}(x, ...)

\method{print}{STACCatalog}(x, ...)

\method{print}{STACCollectionList}(x, n = 10, ...)

\method{print}{STACCollection}(x, ...)

\method{print}{STACItemCollection}(x, n = 10, ..., tail = FALSE)

\method{print}{STACItem}(x, ...)
}
\arguments{
\item{x}{either a \code{RSTACQuery} object expressing a STAC query
criteria or any \code{RSTACDocument}.}

\item{...}{other parameters passed in the functions.}

\item{n}{number of entries to print. Each object has its own rule of
truncation: the \code{STACCollection} objects will print
10 links by default. If the object has less than 20 collections, all
collections will be shown. In \code{STACItemCollection}, 10 features
will be printed by default. To show all entries, use \code{n = Inf}.}

\item{tail}{A \code{logical} value indicating if last features in
STACItemCollection object must be show.}
}
\description{

}
\examples{
\donttest{
# STACItemCollection object
stac_item_collection <-
  stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
  stac_search(collections = "CB4_64_16D_STK-1",
         bbox = c(-47.02148, -12.98314, -42.53906, -17.35063),
         limit = 15) \%>\%
  get_request()

print(stac_item_collection, n = 10)

# STACCollectionList object
stac_collection <-
    stac("https://brazildatacube.dpi.inpe.br/stac/") \%>\%
    collections() \%>\%
    get_request()

print(stac_collection, n = 5)

# RSTACQuery object
obj_rstac <- stac("https://brazildatacube.dpi.inpe.br/stac/")

print(obj_rstac)
}
}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{collections}}
\code{\link{items}}
}
