% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_functions}
\alias{print_functions}
\alias{print.stac}
\alias{print.stac_catalog}
\alias{print.stac_collection}
\alias{print.stac_items}
\alias{print.stac_item}
\title{Printing functions}
\usage{
\method{print}{stac}(x, ...)

\method{print}{stac_catalog}(x, n = 10, ...)

\method{print}{stac_collection}(x, n = 10, ...)

\method{print}{stac_items}(x, n = 5, ...)

\method{print}{stac_item}(x, ...)
}
\arguments{
\item{x}{object STAC to print.}

\item{...}{other parameters passed in the functions.}

\item{n}{number of lines to view on each object. Each object has its own type
of truncation in lines. In the \code{stac_catalog} object, by default, 10
links will be shown. In the \code{stac_collection} object, similarly, 10
links will be shown. In \code{stac_items}, by default, 3 features will be
shown. If you want to show all lines of a rstac object, use `n = Inf`.}
}
\description{
The print functions cover all objects in the rstac package:
\itemize{
  \item \code{\link{stac}}: implements STAC \code{/stac} endpoint.
  \item \code{\link{collections}}: implements \code{/collections}
    and \code{/collections/\{collectionId\}} WFS3 endpoints.
  \item \code{\link{items}}: implements
    \code{/collections/\{collectionId\}/items} and
    \code{/collections/\{collectionId\}/items/\{itemId\}} WFS3 endpoints.
 }

The rstac package objects visualization is based on **Markdown**, a
 lightweight markup language, so you can paste the output into any
 **Markdown** editor for a better visualization.

For printing use the `print()` function directly, since the package has a
 generic implementation for its objects. For console output control, you have
 the option to determine how many items you want to see through the `n` in
 `print` objects parameters, the following objects have the `n` parameter:
 * \code{\link{items}}
 * \code{\link{collections}}
 * \code{\link{stac}}
}
\examples{
\dontrun{

# stac_items object
stac_items <- stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
 search(collections = "MOD13Q1",
        bbox = c(-55.16335, -4.26325, -49.31739, -1.18355),
        limit = 15) \%>\% get_request()

print(stac_items, n = 10)

# stac_catalog object
stac_catalog <-  stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
 collections() \%>\%
 get_request()

print(stac_catalog, n = 5)

# stac object
obj_stac <- stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0")

print(obj_stac)
}
}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{collections}}
\code{\link{items}}
}
\author{
Felipe Carvalho and Rolf Simoes
}
