% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_request}
\alias{get_request}
\alias{post_request}
\title{STAC API request functions}
\usage{
get_request(s, ..., headers = character())

post_request(
  s,
  ...,
  enctype = c("json", "multipart", "form"),
  headers = character()
)
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{collections},
or \code{items} functions.}

\item{...}{other params to be passed to \link[httr]{GET} or
\link[httr]{POST} methods}

\item{headers}{a \code{character} of named arguments to be passed as
HTTP request headers.}

\item{enctype}{a \code{character} informing the request body
Content-Type. Accepted types are \code{'json'} (\code{'application/json'}),
\code{'form'} (\code{'application/x-www-form-urlencoded'}),
and \code{'multipart'} (\code{'multipart/form-data'}). Defaults to
\code{'json'}.}
}
\value{
Either a \code{stac_catalog}, \code{stac_collection},
 \code{stac_collection_list}, \code{stac_item_collection} or \code{stac_item}
 object depending on the subclass and search fields parameters of \code{s}
 argument.
}
\description{
The \code{get_request} is function that makes HTTP GET
requests to STAC web services, retrieves, and parse the data.

The \code{post_request} is function that makes HTTP POST
requests to STAC web services, retrieves, and parse the data.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.1",
      force_version = "0.8.1") \%>\%
 get_request()

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.1",
     force_version = "0.8.1") \%>\%
 stac_search(collections = "MOD13Q1") \%>\%
 post_request()
}
}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{collections}}
\code{\link{items}}
}
