% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac-package.R
\docType{package}
\name{rstac}
\alias{rstac}
\alias{_PACKAGE}
\title{rstac: R client library for SpatioTemporal Asset Catalog (rstac)}
\description{
...
}
\author{
\strong{Maintainer}: Rolf Simoes \email{rolfsimoes@inpe.br} (\href{https://orcid.org/0000-0003-0953-4132}{ORCID})

Authors:
\itemize{
  \item Felipe Souza
}

}
