% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{sign_planetary_computer}
\alias{sign_planetary_computer}
\title{Signature in hrefs provided by the STAC from Microsoft's Planetary
Computer.}
\usage{
sign_planetary_computer(url, env, ...)
}
\arguments{
\item{url}{a \code{character} with the reference href link provided by
objects \code{STACItem} or \code{STACItemCollection}.}

\item{env}{a \code{enviroment} object created to store local variables of the
package.}

\item{...}{additional parameters can be supplied to the \code{GET} function
of the \code{httr} package.}
}
\value{
a \code{character} with the href signed by the token.
}
\description{
To perform the signing of the hrefs a request is sent to
Planetary Computer servers and the returned content corresponds to the
token that will be used in the href.
}
