% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{stac}
\alias{stac}
\title{stac functions}
\usage{
stac(url, .headers = list())
}
\arguments{
\item{url}{A \code{character} informing the base url of a
STAC web service.}

\item{.headers}{A \code{list} of named arguments to be passed as
http request headers.}
}
\value{
A STAC Catalog definition.
}
\description{
This function implements \code{/stac} API
endpoint (v0.8.0). It retrieves the root STAC Catalog or STAC Collection
that is the entry point to access any data published in STAC web service.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0")
}

}
\author{
Rolf Simoes
}
