% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_items.R
\name{stac_items}
\alias{stac_items}
\title{Endpoint functions}
\usage{
stac_items(url, collection_id, item_id, datetime, bbox, limit, ...)
}
\arguments{
\item{url}{a \code{character} informing the base url of a
STAC web service.}

\item{collection_id}{a \code{character} with a collection id to retrieve
collection details.}

\item{item_id}{a \code{character} with item id to be fetched.
Only works if the \code{collection_id} is informed. This is equivalent to
the endpoint \code{/collections/\{collectionId\}/items/\{itemId\}}.}

\item{datetime}{either a date-time or an interval.
Date and time strings needs to conform RFC 3339. Intervals are
expressed by separating two date-time strings by \code{'/'} character.
Open intervals are expressed by using \code{'..'} in place of date-time.

Examples:
\itemize{
  \item A date-time: \code{"2018-02-12T23:20:50Z"}
  \item A closed interval: \code{"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"}
  \item Open intervals: \code{"2018-02-12T00:00:00Z/.."} or
    \code{"../2018-03-18T12:31:12Z"}
}

Only features that have a \code{datetime} property that intersects
the interval or date-time informed in \code{datetime} are selected.}

\item{bbox}{only features that have a geometry that intersects the
bounding box are selected. The bounding box is provided as four or six
numbers, depending on whether the coordinate reference system includes a
vertical axis (elevation or depth):
\itemize{
  \item Lower left corner, coordinate axis 1
  \item Lower left corner, coordinate axis 2
  \item Lower left corner, coordinate axis 3 (optional)
  \item Upper right corner, coordinate axis 1
  \item Upper right corner, coordinate axis 2
  \item Upper right corner, coordinate axis 3 (optional)
}

The coordinate reference system of the values is WGS84
longitude/latitude (\url{http://www.opengis.net/def/crs/OGC/1.3/CRS84}).
The values are in most cases the sequence of minimum longitude,
minimum latitude, maximum longitude and maximum latitude. However,
in cases where the box spans the antimeridian the first value
(west-most box edge) is larger than the third value
(east-most box edge).}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed it defaults to the service implementation.}

\item{...}{filter parameters. Accept the same filter parameters
of \code{\link{stac_search}} function.}
}
\value{
A \code{stac} object containing all request parameters to be
provided to \code{stac_request}.
}
\description{
The \code{stac_items} function implements WFS3
\code{/collections/\{collectionId\}/items}, and
\code{/collections/\{collectionId\}/items/\{itemId\}}
endpoints (v0.8.0).

Each endpoint retrieves specific STAC objects:
\itemize{
  \item \code{/collections/\{collectionId\}/items}: Returns a STAC Items
    collection (GeoJSON)
  \item \code{/collections/\{collectionId\}/items/\{itemId\}}: Returns a
    STAC Item (GeoJSON Feature)
}

The endpoint \code{/collections/\{collectionId\}/items} accepts the same
filters parameters of \code{\link{stac_search}} function.
}
\examples{
\dontrun{

stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "MOD13Q1",
           bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) \%>\%
get_request()

y <- stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "MOD13Q1",
           item_id = "MOD13Q1.A2019241.h13v09.006.2019262164754") \%>\%
get_request()
}

}
\seealso{
\code{\link{get_request}},  \code{\link{post_request}},
 \code{\link{stac_collections}}
}
\author{
Rolf Simoes
}
