% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{stac_request}
\alias{stac_request}
\title{STAC functions}
\usage{
stac_request(
  s,
  limit = 10,
  method = c("get", "post"),
  post_enctype = c("application/json", "application/x-www-form-urlencoded",
    "multipart/form-data"),
  headers = list()
)
}
\arguments{
\item{s}{A \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{limit}{An \code{integer} defining the maximum number of results
to return. If \code{NULL} it defaults to the service implementation.
Defaults to 10.}

\item{method}{A \code{character} value informing the HTTP method to be
used for this request. Accepted methods are \code{'get'} or \code{'post'}.}

\item{post_enctype}{A \code{character} informing the request body
Content-Type. Accepted types \code{'application/json'},
\code{'application/x-www-form-urlencoded'}, and
\code{'multipart/form-data'}}

\item{headers}{A \code{list} of named arguments to be passed as
HTTP request headers.}
}
\value{
Either a \code{stac_collection} or a \code{stac_items} object
depending of the \code{s} parameter.
}
\description{
The \code{stac_request} is function that makes HTTP
requests to STAC web services, retrieves, and parse the data.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    stac_request(limit = 100)
}

}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{stac_collections}}
\code{\link{stac_items}}
}
\author{
Rolf Simoes
}
