% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{stac_request}
\alias{stac_request}
\title{STAC functions}
\usage{
stac_request(stac, method = c("get", "post"), .headers = list())
}
\arguments{
\item{stac}{A \code{stac} object expressing a STAC search criteria.}

\item{method}{A \code{character} value informing the HTTP method to be
used for this request. Accepted methods are \code{'get'} or \code{'post'}.}

\item{.headers}{A \code{list} of named arguments to be passed as
http request headers.}
}
\value{
A \code{stac_catalog} object representing STAC Catalog definition.
}
\description{
This function implements \code{/stac} API
endpoint (v0.8.0). It retrieves the root STAC Catalog or STAC Collection
that is the entry point to access any data published in a STAC web service.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0")
}

}
\seealso{
\code{\link{stac}}, \code{\link{stac_search}},
\code{\link{stac_collections}}, and \code{\link{stac_items}}
}
\author{
Rolf Simoes
}
