% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_request.R
\name{get_request}
\alias{get_request}
\alias{post_request}
\title{STAC functions}
\usage{
get_request(s, headers = c())

post_request(s, enctype = c("json", "multipart", "form"), headers = c())
}
\arguments{
\item{s}{a \code{stac} object expressing a STAC search criteria
provided by \code{stac}, \code{stac_search}, \code{stac_collections},
or \code{stac_items} functions.}

\item{headers}{a \code{character} of named arguments to be passed as
HTTP request headers.}

\item{enctype}{a \code{character} informing the request body
Content-Type. Accepted types are \code{'json'} (\code{'application/json'}),
\code{'form'} (\code{'application/x-www-form-urlencoded'}),
and \code{'multipart'} (\code{'multipart/form-data'}). Defaults to
\code{'json'}.}
}
\value{
Either a \code{stac_collection} or a \code{stac_items} object
depending of the \code{s} parameter.
}
\description{
The \code{get_request} is function that makes HTTP GET
requests to STAC web services, retrieves, and parse the data.

The \code{post_request} is function that makes HTTP POST
requests to STAC web services, retrieves, and parse the data.
}
\examples{
\dontrun{

stac("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0") \%>\%
    get_request()
}

\dontrun{

stac_search("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
            collections = "MOD13Q1") \%>\%
     post_request(enctype = "json")
}

}
\seealso{
\code{\link{stac}} \code{\link{stac_search}} \code{\link{stac_collections}}
\code{\link{stac_items}}
}
\author{
Rolf Simoes and Felipe Carvalho
}
